backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "krilin-ai-backend"
version = "0.1.0"
description = "FastAPI backend for Krilin AI personal assistant"
authors = ["Krilin AI Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pydantic = "^2.10.0"
pydantic-ai-slim = {extras = ["openai", "anthropic", "groq"], version = "^1.0.12"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
redis = "^5.2.0"
celery = "^5.4.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2.1"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
pydantic-settings = "^2.7.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
openai = "^1.58.1"
email-validator = "^2.2.0"
sqladmin = "^0.20.0"
itsdangerous = "^2.2.0"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-api-python-client = "^2.154.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
pre-commit = "^4.0.1"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true